+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformBasePlugin.kt
@@ -5,7 +5,7 @@ package org.jetbrains.intellij.platform.gradle.plugins.project
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.plugins.JavaPlugin
+import org.gradle.api.plugins.JavaLibraryPlugin
 import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.get
 import org.gradle.plugins.ide.idea.IdeaPlugin
@@ -43,7 +43,10 @@ abstract class IntelliJPlatformBasePlugin : Plugin<Project> {
         checkGradleVersion()
 
         with(project.plugins) {
-            apply(JavaPlugin::class)
+            // https://docs.gradle.org/current/userguide/java_plugin.html
+            // https://docs.gradle.org/current/userguide/java_library_plugin.html
+            apply(JavaLibraryPlugin::class)
+            // https://docs.gradle.org/current/userguide/idea_plugin.html
             apply(IdeaPlugin::class)
         }